+++ src/core/observations/deformable_objects/landmarks/landmark.py
@@ -21,6 +21,7 @@ class Landmark:
 
     # Constructor.
     def __init__(self, dimension, tensor_scalar_type):
+        assert dimension is not None, 'dimension can not be None'
         self.dimension = dimension
         self.tensor_scalar_type = tensor_scalar_type
         self.type = 'Landmark'