+++ src/in_out/deformable_object_reader.py
@@ -16,6 +16,7 @@ from in_out.image_functions import normalize_image_intensities
 logger = logging.getLogger(__name__)
 logging.getLogger('PIL').setLevel(logging.WARNING)
 
+
 class DeformableObjectReader:
     """
     Creates PyDeformetrica objects from specified filename and object type.
@@ -25,66 +26,73 @@ class DeformableObjectReader:
     connectivity_degrees = {'LINES': 2, 'POLYGONS': 3}
 
     # Create a PyDeformetrica object from specified filename and object type.
-    def create_object(self, object_filename, object_type, dimension, tensor_scalar_type):
+    @staticmethod
+    def create_object(object_filename, object_type, tensor_scalar_type, dimension=None):
 
-        if object_type.lower() in ['SurfaceMesh'.lower(), 'PolyLine'.lower(),
-                                   'PointCloud'.lower(), 'Landmark'.lower()]:
+        if object_type.lower() in ['SurfaceMesh'.lower(), 'PolyLine'.lower(), 'PointCloud'.lower(), 'Landmark'.lower()]:
 
             if object_type.lower() == 'SurfaceMesh'.lower():
+                points, dimension, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
                 out_object = SurfaceMesh(dimension, tensor_scalar_type)
-                points, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
                 out_object.set_points(points)
                 out_object.set_connectivity(connectivity)
 
             elif object_type.lower() == 'PolyLine'.lower():
+                points, dimension, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
                 out_object = PolyLine(dimension, tensor_scalar_type)
-                points, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
                 out_object.set_points(points)
                 out_object.set_connectivity(connectivity)
 
             elif object_type.lower() == 'PointCloud'.lower():
-                out_object = PointCloud(dimension, tensor_scalar_type)
                 try:
-                    points, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
+                    points, dimension, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
+                    out_object = PointCloud(dimension, tensor_scalar_type)
                     out_object.set_points(points)
                     out_object.set_connectivity(connectivity)
                 except KeyError:
-                    points = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=False)
+                    points, dimension = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=False)
+                    out_object = PointCloud(dimension, tensor_scalar_type)
                     out_object.set_points(points)
 
             elif object_type.lower() == 'Landmark'.lower():
-                out_object = Landmark(dimension, tensor_scalar_type)
                 try:
-                    points, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
+                    points, dimension, connectivity = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=True)
+                    out_object = Landmark(dimension, tensor_scalar_type)
                     out_object.set_points(points)
                     out_object.set_connectivity(connectivity)
                 except KeyError:
-                    points = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=False)
+                    points, dimension = DeformableObjectReader.read_vtk_file(object_filename, dimension, extract_connectivity=False)
+                    out_object = Landmark(dimension, tensor_scalar_type)
                     out_object.set_points(points)
+            else:
+                raise TypeError('Object type ' + object_type + ' was not recognized.')
 
             out_object.update()
 
         elif object_type.lower() == 'Image'.lower():
             if object_filename.find(".png") > 0:
                 img_data = np.array(pimg.open(object_filename))
+                dimension = len(img_data.shape)
                 img_affine = np.eye(dimension + 1)
                 if len(img_data.shape) > 2:
-                    msg = 'Multi-channel images are not managed (yet). Defaulting to the first channel.'
-                    warnings.warn(msg)
+                    warnings.warn('Multi-channel images are not managed (yet). Defaulting to the first channel.')
+                    dimension = 2
                     img_data = img_data[:, :, 0]
 
             elif object_filename.find(".npy") > 0:
                 img_data = np.load(object_filename)
+                dimension = len(img_data.shape)
                 img_affine = np.eye(dimension + 1)
 
             elif object_filename.find(".nii") > 0:
                 img = nib.load(object_filename)
                 img_data = img.get_data()
+                dimension = len(img_data.shape)
                 img_affine = img.affine
                 assert len(img_data.shape) == 3, "Multi-channel images not available (yet!)."
 
             else:
-                raise ValueError('Unknown image extension for file: %s' % object_filename)
+                raise TypeError('Unknown image extension for file: %s' % object_filename)
 
             # Rescaling between 0. and 1.
             img_data, img_data_dtype = normalize_image_intensities(img_data)
@@ -95,8 +103,7 @@ class DeformableObjectReader:
 
             dimension_image = len(img_data.shape)
             if dimension_image != dimension:
-                logger.warning('I am reading a {}d image but the dimension is set to {}'
-                               .format(dimension_image, dimension))
+                logger.warning('I am reading a {}d image but the dimension is set to {}'.format(dimension_image, dimension))
 
             out_object.update()
 
@@ -106,7 +113,7 @@ class DeformableObjectReader:
         return out_object
 
     @staticmethod
-    def read_vtk_file(filename, dimension, extract_connectivity=False):
+    def read_vtk_file(filename, dimension=None, extract_connectivity=False):
         """
         Routine to read  vtk files
         Probably needs new case management
@@ -122,6 +129,13 @@ class DeformableObjectReader:
         nb_points = int(fifth_line[1])
         points = []
         line_start_connectivity = None
+        connectivity_type = nb_faces = nb_vertices_in_faces = None
+
+        if dimension is None:
+            dimension = DeformableObjectReader.__detect_dimension(content)
+
+        assert isinstance(dimension, int)
+        logger.debug('Using dimension ' + str(dimension) + ' for file ' + filename)
 
         # Reading the points:
         for i in range(5, len(content)):
@@ -142,6 +156,14 @@ class DeformableObjectReader:
 
         # Reading the connectivity, if needed.
         if extract_connectivity:
+            # Error checking
+            if connectivity_type is None:
+                RuntimeError('Could not determine connectivity type.')
+            if nb_faces is None:
+                RuntimeError('Could not determine number of faces type.')
+            if nb_vertices_in_faces is None:
+                RuntimeError('Could not determine number of vertices type.')
+
             if line_start_connectivity is None:
                 raise KeyError('Could not read the connectivity for the given vtk file')
 
@@ -152,14 +174,12 @@ class DeformableObjectReader:
                 number_vertices_in_line = int(line[0])
 
                 if connectivity_type == 'POLYGONS':
-                    assert number_vertices_in_line == 3, 'Invalid connectivity: ' \
-                                                         'deformetrica only handles triangles for now.'
+                    assert number_vertices_in_line == 3, 'Invalid connectivity: deformetrica only handles triangles for now.'
                     connectivity.append([int(elt) for elt in line[1:]])
-
                 elif connectivity_type == 'LINES':
                     assert number_vertices_in_line >= 2, 'Should not happen.'
-                    for i in range(1, number_vertices_in_line):
-                        connectivity.append([int(line[i]), int(line[i+1])])
+                    for j in range(1, number_vertices_in_line):
+                        connectivity.append([int(line[j]), int(line[j+1])])
 
             connectivity = np.array(connectivity)
 
@@ -168,6 +188,33 @@ class DeformableObjectReader:
                 assert len(connectivity) == nb_faces, 'Found an unexpected number of faces.'
                 assert len(connectivity) * 4 == nb_vertices_in_faces
 
-            return points, connectivity
+            return points, dimension, connectivity
 
-        return points
+        return points, dimension
+
+    @staticmethod
+    def __detect_dimension(content, nb_lines_to_check=2):
+        """
+        Try to determine dimension from VTK file: check last element in first nb_lines_to_check points to see if filled with 0.00000, if so 2D else 3D
+        :param content:     content to check
+        :param nb_lines_to_check:   number of lines to check
+        :return:    detected dimension
+        """
+        assert nb_lines_to_check > 0, 'You must check at least 1 line'
+
+        dimension = None
+
+        for i in range(5, 5+nb_lines_to_check-1):
+            line_elements = content[i].split(' ')
+            if float(line_elements[2]) == 0.:
+                if dimension is not None and dimension == 3:
+                    raise RuntimeError('Could not automatically determine data dimension. Please manually specify value.')
+                dimension = 2
+            elif float(line_elements[2]) != 0.:
+                if dimension is not None and dimension == 2:
+                    raise RuntimeError('Could not automatically determine data dimension. Please manually specify value.')
+                dimension = 3
+            else:
+                raise RuntimeError('Could not automatically determine data dimension. Please manually specify value.')
+
+        return dimension