+++ tests/functional_tests/functional_test.py
@@ -110,9 +110,10 @@ class FunctionalTest(unittest.TestCase):
         actual = read_3D_array(path_to_actual_txt_file)
         self._compare_numpy_arrays(expected, actual)
 
-    def _compare_vtk_files(self, path_to_expected_vtk_file, path_to_actual_vtk_file, dimension=2):
-        expected = DeformableObjectReader.read_vtk_file(path_to_expected_vtk_file, dimension)
-        actual = DeformableObjectReader.read_vtk_file(path_to_actual_vtk_file, dimension)
+    def _compare_vtk_files(self, path_to_expected_vtk_file, path_to_actual_vtk_file):
+        expected, expected_dimension = DeformableObjectReader.read_vtk_file(path_to_expected_vtk_file)
+        actual, dimension = DeformableObjectReader.read_vtk_file(path_to_actual_vtk_file)
+        self.assertEqual(expected_dimension, dimension)
         self._compare_numpy_arrays(expected, actual)
 
     def _compare_png_files(self, path_to_expected_png_file, path_to_actual_png_file):