+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- new API: deformetrica can now be instantiated from python
+- New API: deformetrica can now be instantiated from python
+- Automatic dimension detection from input file
 
 
 ## [4.0.0] - 2018-06-14